Convert Utils/Registry to TS
