+++ packages/utils/src/Registry/Registry.test.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+import ReducerRegistry from '../ReducerRegistry';
 import registry, { getRegistry } from './Registry';
 
 it('getRegistry', () => {
@@ -8,11 +10,19 @@ it('getRegistry', () => {
 });
 
 it('registry decorator', () => {
-  const App = function () {};
+  class App extends React.Component {
+    render() {
+      return 'app';
+    }
+  }
 
   registry()(App);
 
-  const instance = new App();
+  interface ObjectWithRegistry extends App {
+    getRegistry: () => ReducerRegistry<any> | void;
+  }
+
+  const instance = new App({}) as ObjectWithRegistry;
 
   expect(instance.getRegistry()).toMatchObject({
     reducers: { routerData: expect.any(Function) },
